Extract dataclasses into __init__
